It builds, therefore, it is.
The current problems with this build script are that it generates an ISO
that is 3.7 GB.  I would like to attempt it with Xubuntu, Debian, and
also the mini debian boots to see if I can get it to work better off of
them.  I might have to hook the Debian installer for that, though, but
not a problem.

Some notes on the things I noticed that are duplicated:

1) 2 kernels (probably through upgrade process)
2) 2 JREs
3) 3 versions of Python

So while `apt-get -y purge` commands would potentially clean these up,
they just become a game of Whack a Mole really quickly with other
packages that get added by whatever base distro you use and becomes
annoying really quickly (I've done it before, it's not worth the
effort).

I'll test these later when I have more time and try to fix up the build
a bit.  Hopefully I can produce a nice decent sized ISO for DAVIX.

Also to note, this boots, but requires (at the moment) a base image that
already boots to a desktop.  So for instance, an installer image like
Ubuntu Server (currently) doesn't work since it just boots directly to
the installer menu.  I can bypass this, but at the moment I'm more
concerned with figuring out which base image to use.